# apps/api/openapi/spec.yaml
openapi: 3.0.0
info:
  title: lomi.africa Payments API
  version: 1.0.0
  description: API of lomi. payment orchestration platform
servers:
  - url: https://api.lomi.africa/v1
    
# API Schema to Database Table Mapping:
# - Merchant (API) -> merchants (DB)
# - Product (API) -> merchant_products (DB)
# - SubscriptionPlan (API) -> subscription_plans (DB)
# - Subscription (API) -> merchant_subscriptions (DB)
# - Transaction (API) -> transactions (DB)
# - Customer (API) -> customers (DB)
# - PaymentLink (API) -> payment_links (DB)
# - Refund (API) -> refunds (DB)
# - Webhook (API) -> webhooks (DB)
# - CheckoutSession (API) -> checkout_sessions (DB)
# - MerchantAccount (API) -> merchant_accounts (DB)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    Refund:
      type: object
      properties:
        refund_id:
          type: string
          format: uuid
        transaction_id:
          type: string
          format: uuid
        amount:
          type: number
        refunded_amount:
          type: number
        fee_amount:
          type: number
        status:
          $ref: '#/components/schemas/RefundStatus'
        reason:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - refund_id
        - transaction_id
        - amount
        - refunded_amount
        - fee_amount
        - status
    CreateRefund:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0
        reason:
          type: string
        metadata:
          type: object
      required:
        - transaction_id
        - amount
...
    RefundStatus:
      type: string
      enum:
        - pending
        - completed
        - failed
...
security:
  - ApiKeyAuth: []
paths:
...
  /refunds:
    post:
      summary: Create a refund
      description: Initiate a refund for a transaction
      operationId: createRefund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRefund'
      responses:
        '201':
          description: Refund created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /refunds/{refund_id}:
    get:
      summary: Get refund details
      description: Get details of a specific refund
      operationId: getRefund
      parameters:
        - name: refund_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Refund details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update refund status
      description: Update the status of a refund
      operationId: updateRefund
      parameters:
        - name: refund_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [completed, failed]
                metadata:
                  type: object
              required:
                - status
      responses:
        '200':
          description: Refund updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' 
...